requirements.txt
# packages used in app
requests
python-dotenv
random

# packages used for testing:
pytest